Interpolate quantiles and return a cumulative distribution function

Interpolate quantiles and return a cumulative distribution function

interpolate_quantiles(object, ...)

# S3 method for qs
interpolate_quantiles(
  object,
  newdata = NULL,
  grid = seq_quant(0, 1, by = 0.01),
  parallel = T,
  ncores = getCores(),
  row_thresh = 20,
  tails = "gaussian",
  distn = "q",
  ...
)

# S3 method for matrix
interpolate_quantiles(
  object,
  alphas,
  grid = seq_quant(0, 1, by = 0.01),
  parallel = T,
  ncores = getCores(),
  row_thresh = 100,
  tails = "gaussian",
  distn = "q",
  ...
)

Arguments

object
either matrix of quantiles or a fitted model with a relevant method to dispatch

...
other arguments passed to interpolate quantiles

newdata
New data to use when calculating quantiles to interpolate

grid
grid along which to evaluate the CDF

parallel
whether to work in parallel

ncores
number of cores to use

row_thresh
required minimum number of observations to use parallel processing

tails
what distribution to use when fitting the tails, either "gaussian" or "exponential"

distn
what to return, q for quantile, c for cdf, p for pdf

alphas
values those quantiles are fit to