Compute quantile regressions via quantile spacings

Compute quantile regressions via quantile spacings

qs(
  formula,
  data = NULL,
  quantiles = c(0.9, 0.75, 0.5, 0.25, 0.1),
  baseline_quantile = 0.5,
  calc_se = T,
  se_method = "subsample",
  weight_vec = NULL,
  subsample_percent = 0.2,
  algorithm = "sfn",
  cluster_formula = NULL,
  num_bs = 100,
  parallel = TRUE,
  num_cores = getCores(),
  trunc = T,
  small = NULL,
  seed = NULL,
  draw_weights = NULL,
  sampling_method = NULL,
  output_quantiles = T,
  calc_avg_me = F,
  lambda = NULL,
  ...
)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula)

quantiles
vector of quantiles to be estimated

baseline_quantile
baseline quantile to measure spacings from (defaults to 0.5)

calc_se
boolean, whether or not to calculate standard errors

se_method
Method to use for standard errors, either "weighted_bootstrap", "subsample", "bootstrap" or "resample_qs" along with a specified subsampling method and subsample percent.

weight_vec
vector of weights for weighted quantile regression

subsample_percent
percent to subsample for standard error calculations

algorithm
What algorithm to use for fitting underlying regressions. Either one of "sfn", "br", "lasso" or a function name which estimates quantiles

cluster_formula
formula (e.g. ~X1 + X2) giving the clustering formula

num_bs
number of bootstrap draws

parallel
whether to run bootstrap in parallel

num_cores
number of cores to use (defaults to setting from `getOption(mc.cores)`)

trunc
whether to truncate small values

small
level of "small" values to guarentee numerical stability

seed
what seed to use for replicable RNG

draw_weights
whether to use random exponential weights for bootstrap

sampling_method
One of "leaveRows", "subsampleRows", or "bootstrapRows". leaveRows doesn't resample rows at all. subsampleRows samples without replacement given some percentage of the data (specified via subsample_percent), and boostrapRows samples with replacement.

output_quantiles
whether to save fitted quantiles as part of the function output

calc_avg_me
whether to return average marginal effects as part of the fitted object

lambda
optional penalty parameter, ignored except for penalized quantile regressions

...
additional arguments, ignored for now