Runs quantile regression on residuals of the model (calculates spaces around jstar quantile)

Runs quantile regression on residuals of the model (calculates spaces around jstar quantile)

regressResiduals(
  reg_spec_data,
  ehat,
  ind_hat,
  tau,
  trunc,
  small,
  weights,
  algorithm = "rq.fit.sfn_start_val",
  ...
)

Arguments

reg_spec_data

result of ensureSpecRank function; regression matrix with full rank

ehat

current residuals; subset of which to be used as dependent column

ind_hat

column vector indicating which rows to be used in quantile regression

tau

estimated quantile

trunc

Boolean value; if true, replace those dependent values less than small with small itself; else, only use rows with residuals greater than small

small

Value used with trunc; values less than small 'blow up' too greatly when logged

weights

vector of optional weights

algorithm

The name of a function which will estimate a quantile regression. Defaults to rq.fit.sfn_start_val. Must be a string, as it is passed to do.call

...

other arguments to the function specified by the algorithm argument

Value

List of estimated coefficients, warnings, iterations, and controls as in standard quantile regression function