Control standard_errors parameters

Control standard_errors parameters

se_control(
  se_method = "subsample",
  subsample_percent = 0.2,
  num_bs = 100,
  draw_weights = F,
  sampling_method = "subsampleRows",
  parallel_thresh = 1e+05,
  ...
)

Arguments

se_method
Method to use for standard errors, either "weighted_bootstrap", "subsample", "bootstrap" or "rq_sample" (the last one is fully customized per the other se_control parameters) along with a specified subsampling method and subsample percent. If specifying "custom", must also specify subsampling_percent and draw_weights. If you specify "subsample", subsampling percent defaults to 0.2, but can be changed.

subsample_percent
A number between 0 and one, specifying the percent of the data to subsample for standard error calculations

num_bs
Number of bootstrap iterations to use, defaults to 100.

draw_weights
Whether to use random exponential weights for bootstrap, either TRUE or FALSE

sampling_method
One of "leaveRows", "subsampleRows", or "bootstrapRows". leaveRows doesn't resample rows at all. subsampleRows samples without replacement given some percentage of the data (specified via subsample_percent), and bootstrapRows samples with replacement.

parallel_thresh
threshold for when to use parallel, based on nrow(X) * num_bs. To always use parallel, set to 0. Designed to avoid slow overhead when problem is relatively small.

...
Other arguments, ignored for now

Details

se_control control parameters to pass to the control arguments of quantreg_spacing, the lower-level function called by standard_errors.